Application/requirements.txt
opencv-python~=4.5.4.60
mediapipe~=0.8.9


sklearn~=0.0
scikit-learn~=1.0.1
pyngrok~=5.1.0
pandas~=1.3.4
numpy~=1.21.4
paho-mqtt~=1.6.1